feat: implement production-level outbound adapters and zero compilation errors
- Implement production-level HTTP proxy outbound adapter with TLS support
- Implement production-level SOCKS5 outbound adapter with UDP ASSOCIATE
- Implement production-level Shadowsocks outbound adapter with AES-256-GCM and ChaCha20-Poly1305
- Add ResolveMode clap ValueEnum support for CLI integration
- Fix IPv6 stability API usage with stable implementation
- Resolve all Arc type issues and memory management
- Add comprehensive error handling and metrics support
- Achieve zero compilation errors across all crates with all features enabled

+++ crates/sb-core/src/outbound/udp_balancer.rs
@@ -167,7 +167,7 @@ pub async fn send_balanced(
         // base=200ms, factor=2, cap=5s
         let base = 200u64;
         let cap = 5000u64;
-        let shift = (fails.min(16)) as u32;
+        let shift = fails.min(16);
         let mul = 1u64 << shift; // 2^fails
         Duration::from_millis((base.saturating_mul(mul)).min(cap))
     }
